cleanup: separate u/a-law
